+++ ariadne/asgi.py
@@ -1,5 +1,6 @@
 import asyncio
 import json
+from dataclasses import dataclass
 from inspect import isawaitable
 from typing import (
     Any,
@@ -38,7 +39,6 @@ from .types import (
     ValidationRules,
 )
 
-
 GQL_CONNECTION_INIT = "connection_init"  # Client -> Server
 GQL_CONNECTION_ACK = "connection_ack"  # Server -> Client
 GQL_CONNECTION_ERROR = "connection_error"  # Server -> Client
@@ -73,8 +73,17 @@ Middlewares = Union[
 ]
 
 
-OnConnect = Callable[[WebSocket, Any], None]
-OnDisconnect = Callable[[WebSocket], None]
+@dataclass
+class Operation:
+    id: str
+    name: Optional[str]
+    generator: AsyncGenerator
+
+
+OnConnect = Callable[[WebSocket, Any], Any]
+OnDisconnect = Callable[[WebSocket], Any]
+OnOperation = Callable[[WebSocket, Operation], Any]
+OnComplete = Callable[[WebSocket, Operation], Any]
 
 
 class GraphQL:
@@ -86,6 +95,8 @@ class GraphQL:
         root_value: Optional[RootValue] = None,
         on_connect: Optional[OnConnect] = None,
         on_disconnect: Optional[OnDisconnect] = None,
+        on_operation: Optional[OnOperation] = None,
+        on_complete: Optional[OnComplete] = None,
         validation_rules: Optional[ValidationRules] = None,
         debug: bool = False,
         introspection: bool = True,
@@ -99,6 +110,8 @@ class GraphQL:
         self.root_value = root_value
         self.on_connect = on_connect
         self.on_disconnect = on_disconnect
+        self.on_operation = on_operation
+        self.on_complete = on_complete
         self.validation_rules = validation_rules
         self.debug = debug
         self.introspection = introspection
@@ -268,7 +281,7 @@ class GraphQL:
         return Response(status_code=405, headers=allow_header)
 
     async def websocket_server(self, websocket: WebSocket) -> None:
-        subscriptions: Dict[str, AsyncGenerator] = {}
+        operations: Dict[str, Operation] = {}
         await websocket.accept("graphql-ws")
         try:
             while WebSocketState.DISCONNECTED not in (
@@ -276,18 +289,29 @@ class GraphQL:
                 websocket.application_state,
             ):
                 message = await websocket.receive_json()
-                await self.handle_websocket_message(message, websocket, subscriptions)
+                await self.handle_websocket_message(message, websocket, operations)
         except WebSocketDisconnect:
             pass
         finally:
-            for subscription in subscriptions.values():
-                await subscription.aclose()
+            for operation_id, operation in list(operations.items()):
+                await self.stop_websocket_operation(websocket, operation)
+                del operations[operation_id]
+
+            try:
+                if self.on_disconnect:
+                    result = self.on_disconnect(websocket)
+                    if result and isawaitable(result):
+                        await result
+            except Exception as error:
+                if not isinstance(error, GraphQLError):
+                    error = GraphQLError(str(error), original_error=error)
+                log_error(error, self.logger)
 
     async def handle_websocket_message(
         self,
         message: dict,
         websocket: WebSocket,
-        subscriptions: Dict[str, AsyncGenerator],
+        operations: Dict[str, Operation],
     ):
         operation_id = cast(str, message.get("id"))
         message_type = cast(str, message.get("type"))
@@ -302,13 +326,13 @@ class GraphQL:
                 websocket,
             )
         elif message_type == GQL_START:
-            await self.start_websocket_subscription(
-                message.get("payload"), operation_id, websocket, subscriptions
+            await self.start_websocket_operation(
+                message.get("payload"), operation_id, websocket, operations
             )
         elif message_type == GQL_STOP:
-            if operation_id in subscriptions:
-                await subscriptions[operation_id].aclose()
-                del subscriptions[operation_id]
+            if operation_id in operations:
+                await self.stop_websocket_operation(websocket, operations[operation_id])
+                del operations[operation_id]
 
     async def handle_websocket_connection_init_message(
         self,
@@ -342,16 +366,6 @@ class GraphQL:
     ):
         await websocket.close()
 
-        try:
-            if self.on_disconnect:
-                result = self.on_disconnect(websocket)
-                if result and isawaitable(result):
-                    await result
-        except Exception as error:
-            if not isinstance(error, GraphQLError):
-                error = GraphQLError(str(error), original_error=error)
-            log_error(error, self.logger)
-
     async def keep_websocket_alive(self, websocket: WebSocket):
         if not self.keepalive:
             return
@@ -362,12 +376,12 @@ class GraphQL:
                 return
             await asyncio.sleep(self.keepalive)
 
-    async def start_websocket_subscription(
+    async def start_websocket_operation(
         self,
         data: Any,
         operation_id: str,
         websocket: WebSocket,
-        subscriptions: Dict[str, AsyncGenerator],
+        operations: Dict[str, Operation],
     ):
         context_value = await self.get_context_for_request(websocket)
 
@@ -389,10 +403,40 @@ class GraphQL:
             )
         else:
             results = cast(AsyncGenerator, results)
-            subscriptions[operation_id] = results
+            operations[operation_id] = Operation(
+                id=operation_id,
+                name=data.get("operationName"),
+                generator=results,
+            )
+
+            if self.on_operation:
+                try:
+                    result = self.on_operation(websocket, operations[operation_id])
+                    if result and isawaitable(result):
+                        await result
+                except Exception as error:
+                    if not isinstance(error, GraphQLError):
+                        error = GraphQLError(str(error), original_error=error)
+                    log_error(error, self.logger)
+
             asyncio.ensure_future(
                 self.observe_async_results(results, operation_id, websocket)
             )
+
+    async def stop_websocket_operation(
+        self, websocket: WebSocket, operation: Operation
+    ):
+        if self.on_complete:
+            try:
+                result = self.on_complete(websocket, operation)
+                if result and isawaitable(result):
+                    await result
+            except Exception as error:
+                if not isinstance(error, GraphQLError):
+                    error = GraphQLError(str(error), original_error=error)
+                log_error(error, self.logger)
+
+        await operation.generator.aclose()
 
     async def observe_async_results(
         self, results: AsyncGenerator, operation_id: str, websocket: WebSocket