+++ tests/asgi/test_websockets.py
@@ -204,7 +204,217 @@ def test_custom_websocket_connection_error_in_custom_websocket_on_connect_is_han
         assert response["payload"] == {"msg": "Token required", "code": "auth_error"}
 
 
-def test_custom_websocket_on_disconnect_is_called(schema):
+def test_custom_websocket_on_operation_is_called(schema):
+    def on_operation(websocket, operation):
+        assert operation.name == "TestOp"
+        websocket.scope["on_operation"] = True
+
+    app = GraphQL(schema, on_operation=on_operation)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
+        ws.send_json(
+            {
+                "type": GQL_START,
+                "id": "test1",
+                "payload": {
+                    "operationName": "TestOp",
+                    "query": "subscription TestOp { ping }",
+                },
+            }
+        )
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "pong"}
+        ws.send_json({"type": GQL_STOP})
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "ping"}
+        ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+        assert ws.scope["on_operation"] is True
+
+
+def test_custom_websocket_on_operation_is_awaited_if_its_async(schema):
+    async def on_operation(websocket, operation):
+        assert operation.name == "TestOp"
+        websocket.scope["on_operation"] = True
+
+    app = GraphQL(schema, on_operation=on_operation)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
+        ws.send_json(
+            {
+                "type": GQL_START,
+                "id": "test1",
+                "payload": {
+                    "operationName": "TestOp",
+                    "query": "subscription TestOp { ping }",
+                },
+            }
+        )
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "pong"}
+        ws.send_json({"type": GQL_STOP})
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "ping"}
+        ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+        assert ws.scope["on_operation"] is True
+
+
+def test_error_in_custom_websocket_on_operation_is_handled(schema):
+    async def on_operation(websocket, operation):
+        raise ValueError("Oh No!")
+
+    app = GraphQL(schema, on_operation=on_operation)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
+        ws.send_json(
+            {
+                "type": GQL_START,
+                "id": "test1",
+                "payload": {
+                    "operationName": "TestOp",
+                    "query": "subscription TestOp { ping }",
+                },
+            }
+        )
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "pong"}
+        ws.send_json({"type": GQL_STOP})
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "ping"}
+        ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+
+
+def test_custom_websocket_on_complete_is_called(schema):
+    def on_complete(websocket, operation):
+        assert operation.name == "TestOp"
+        websocket.scope["on_complete"] = True
+
+    app = GraphQL(schema, on_complete=on_complete)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
+        ws.send_json(
+            {
+                "type": GQL_START,
+                "id": "test1",
+                "payload": {
+                    "operationName": "TestOp",
+                    "query": "subscription TestOp { ping }",
+                },
+            }
+        )
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "pong"}
+        ws.send_json({"type": GQL_STOP})
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "ping"}
+        ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+        assert "on_complete" not in ws.scope
+
+    assert ws.scope["on_complete"] is True
+
+
+def test_custom_websocket_on_complete_is_awaited_if_its_async(schema):
+    async def on_complete(websocket, operation):
+        assert operation.name == "TestOp"
+        websocket.scope["on_complete"] = True
+
+    app = GraphQL(schema, on_complete=on_complete)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
+        ws.send_json(
+            {
+                "type": GQL_START,
+                "id": "test1",
+                "payload": {
+                    "operationName": "TestOp",
+                    "query": "subscription TestOp { ping }",
+                },
+            }
+        )
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "pong"}
+        ws.send_json({"type": GQL_STOP})
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "ping"}
+        ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+        assert "on_complete" not in ws.scope
+
+    assert ws.scope["on_complete"] is True
+
+
+def test_error_in_custom_websocket_on_complete_is_handled(schema):
+    async def on_complete(websocket, operation):
+        raise ValueError("Oh No!")
+
+    app = GraphQL(schema, on_complete=on_complete)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
+        ws.send_json(
+            {
+                "type": GQL_START,
+                "id": "test1",
+                "payload": {
+                    "operationName": "TestOp",
+                    "query": "subscription TestOp { ping }",
+                },
+            }
+        )
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "pong"}
+        ws.send_json({"type": GQL_STOP})
+        response = ws.receive_json()
+        assert response["type"] == GQL_DATA
+        assert response["id"] == "test1"
+        assert response["payload"]["data"] == {"ping": "ping"}
+        ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+
+
+def test_custom_websocket_on_disconnect_is_called_on_terminate_message(schema):
     def on_disconnect(websocket):
         websocket.scope["on_disconnect"] = True
 
@@ -216,6 +426,22 @@ def test_custom_websocket_on_disconnect_is_called(schema):
         response = ws.receive_json()
         assert response["type"] == GQL_CONNECTION_ACK
         ws.send_json({"type": GQL_CONNECTION_TERMINATE})
+        assert "on_disconnect" not in ws.scope
+
+    assert ws.scope["on_disconnect"] is True
+
+
+def test_custom_websocket_on_disconnect_is_called_on_connection_close(schema):
+    def on_disconnect(websocket):
+        websocket.scope["on_disconnect"] = True
+
+    app = GraphQL(schema, on_disconnect=on_disconnect)
+    client = TestClient(app)
+
+    with client.websocket_connect("/", "graphql-ws") as ws:
+        ws.send_json({"type": GQL_CONNECTION_INIT})
+        response = ws.receive_json()
+        assert response["type"] == GQL_CONNECTION_ACK
         assert "on_disconnect" not in ws.scope
 
     assert ws.scope["on_disconnect"] is True